Add files via upload
House Prices - Advanced Regression Techniques :
The real estate market is a complex and dynamic sector, where accurately predicting house prices is both a crucial and challenging task. Understanding the factors that influence house prices and developing robust predictive models can provide significant value to various stakeholders, including homeowners, real estate agents, investors, and policymakers. The "House Prices - Advanced Regression Techniques" project delves into this intricate domain by employing sophisticated statistical and machine learning methods to forecast the sale prices of houses.

This project is part of a competition hosted by Kaggle, aimed at providing a hands-on opportunity to apply advanced regression techniques to a real-world dataset. Participants are tasked with building models that can predict the final price of each home in Ames, Iowa, based on its features. The dataset includes a rich array of variables, such as the size of the house, the number of bedrooms, the quality of construction, the neighborhood, and many more.

By leveraging advanced regression techniques, this project aims to develop highly accurate and interpretable models that can predict house prices with a high degree of precision. The insights gained from this project not only enhance the predictive capabilities but also deepen the understanding of the housing market dynamics.